🚚 Rename :  category 파일 이동
